Correct format string type mismatches in some debugging statements.
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10217 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ xps/xpscff.c
@@ -602,7 +602,7 @@ xps_post_callback_decode_glyph(gs_font *p42, gs_glyph glyph)
 static int
 xps_post_callback_glyph_name(gs_font *pf, gs_glyph glyph, gs_const_string *pstr)
 {
-    dprintf1("asking for CFF glyph name %d\n", glyph);
+    dprintf1("asking for CFF glyph name %lu\n", (ulong)glyph);
     return -1;
 }
 
@@ -610,7 +610,7 @@ static int
 xps_post_callback_glyph_info(gs_font *font, gs_glyph glyph,
 	const gs_matrix *pmat, int members, gs_glyph_info_t *info)
 {
-    dprintf1("asking for CFF glyph info %d\n", glyph);
+    dprintf1("asking for CFF glyph info %lu\n", (ulong)glyph);
     return -1;
 }
 
@@ -618,7 +618,7 @@ static int
 xps_post_callback_glyph_outline(gs_font *font, int wmode, gs_glyph glyph,
 	const gs_matrix *pmat, gx_path *ppath, double sbw[4])
 {
-    dprintf1("asking for CFF glyph outline %d\n", glyph);
+    dprintf1("asking for CFF glyph outline %lu\n", (ulong)glyph);
     return -1;
 }
